# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        }
    },
    "paths": {
...
        "/nfts/activities": {
            "get": {
                "description": "get the activities with related to the NFT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nfts"
                ],
                "summary": "Get NFT activities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NFT address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.NFTActivityRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/nfts/offers": {
            "get": {
                "description": "get the offers with related to the NFT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nfts"
                ],
                "summary": "Get NFT offers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NFT address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.NFTActivity"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/nfts/{address}": {
            "get": {
                "description": "get detail information about the NFT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "nfts"
                ],
                "summary": "Get NFT detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "NFT address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.NFT"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/activities": {
            "get": {
                "description": "get the activities with related to the wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user activities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Activity types (['listing'])",
                        "name": "activity_types",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ActivityRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/nfts": {
            "get": {
                "description": "get the nfts of the wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user NFTs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.UserNFT"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/offers": {
            "get": {
                "description": "get the offers with related to the wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user offers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wallet address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ActivityRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "types.Activity": {
            "type": "object",
            "properties": {
                "activityType": {
                    "type": "string"
                },
                "buyer": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "martketplaceProgramAddress": {
                    "type": "string"
                },
                "mint": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "seller": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "types.ActivityRes": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Activity"
                    }
                },
                "hasNextPage": {
                    "type": "boolean"
                }
            }
        },
...
        "types.CollectedCollection": {
            "type": "object",
            "properties": {
                "estimatedValue": {
                    "type": "string"
                },
                "floorPrice": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nftsOwned": {
                    "type": "integer"
                }
            }
        },
...
        "types.NFT": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "lastSold": {
                    "type": "string"
                },
                "listingPrice": {
                    "type": "string"
                },
                "listingType": {
                    "type": "string"
                },
                "mintAddress": {
                    "type": "string"
                },
                "moonrankRank": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "projectId": {
                    "type": "string"
                },
                "sellerFeeBasisPoints": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                },
                "tokenStandard": {
                    "type": "string"
                },
                "traits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Trait"
                    }
                },
                "uri": {
                    "type": "string"
                }
            }
        },
        "types.NFTActivity": {
            "type": "object",
            "properties": {
                "activityType": {
                    "type": "string"
                },
                "buyer": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "martketplaceProgramAddress": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "seller": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
        "types.NFTActivityRes": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.NFTActivity"
                    }
                },
                "hasNextPage": {
                    "type": "boolean"
                }
            }
        },
...
        "types.Trait": {
            "type": "object",
            "properties": {
                "rarity": {
                    "type": "number"
                },
                "traitType": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "types.UserNFT": {
            "type": "object",
            "properties": {
                "collections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.CollectedCollection"
                    }
                },
                "nfts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.NFT"
                    }
                }
            }
        }
    }
}